Move workspace package import files
